PHP files for Login/Signup
Please note: these have not been tested! For whatever reason, it wants to be unformatted. Could be because it's not on a server. 
Once information is provided, remove the comments around $db = newsqli("type", "username", "password", "username"); and reformat so it can login to the server and do all the verification / adding stuff.